Refactor the agent config code a bit more
