+++ src/Microsoft.SqlTools.ServiceLayer/Agent/AgentService.cs
@@ -21,13 +21,14 @@ using Microsoft.SqlTools.Utility;
 namespace Microsoft.SqlTools.ServiceLayer.Agent
 {
     /// <summary>
-    /// Main class for Profiler Service functionality
+    /// Main class for Agent Service functionality
     /// </summary>
     public sealed class AgentService
-    {
-        private Dictionary<Guid, JobProperties> jobs = null;
+    {        
         private ConnectionService connectionService = null;
         private static readonly Lazy<AgentService> instance = new Lazy<AgentService>(() => new AgentService());
+
+        private IAgentJobsService agentJobsService = null;
 
         /// <summary>
         /// Construct a new AgentService instance with default parameters
@@ -64,6 +65,24 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
             }
         }
 
+        /// <summary>
+        /// Internal for testing purposes only
+        /// </summary>
+        internal IAgentJobsService AgentJobsServiceInstance
+        {
+            get
+            {
+                if (this.agentJobsService == null)
+                {
+                    this.agentJobsService = new AgentJobsService();
+                }
+                return this.agentJobsService;
+            }
+            set
+            {
+               this. agentJobsService = value;
+            }
+        }
 
         /// <summary>
         /// Service host object for sending/receiving requests/events.
@@ -81,9 +100,15 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
         public void InitializeService(ServiceHost serviceHost)
         {
             this.ServiceHost = serviceHost;
+
+            // Jobs request handlers
             this.ServiceHost.SetRequestHandler(AgentJobsRequest.Type, HandleAgentJobsRequest);
             this.ServiceHost.SetRequestHandler(AgentJobHistoryRequest.Type, HandleJobHistoryRequest);
             this.ServiceHost.SetRequestHandler(AgentJobActionRequest.Type, HandleJobActionRequest);
+
+            // Alerts request handlers
+            this.ServiceHost.SetRequestHandler(AgentAlertsRequest.Type, HandleAgentAlertsRequest);
+
         }
     
         /// <summary>
@@ -91,7 +116,7 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
         /// </summary>
         internal async Task HandleAgentJobsRequest(AgentJobsParams parameters, RequestContext<AgentJobsResult> requestContext)
         {
-            try
+            await Task.Run(async () =>
             {
                 var result = new AgentJobsResult();
                 ConnectionInfo connInfo;
@@ -101,31 +126,21 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
 
                 if (connInfo != null)
                 {
-                    var sqlConnection = ConnectionService.OpenSqlConnection(connInfo);
-                    var serverConnection = new ServerConnection(sqlConnection);
-                    var fetcher = new JobFetcher(serverConnection);
-                    var filter = new JobActivityFilter();
-                    this.jobs = fetcher.FetchJobs(filter);
-
-                    var agentJobs = new List<AgentJobInfo>();
-                    if (this.jobs != null)
+                    List<AgentJobInfo> jobs;
+                    var returnValue = this.AgentJobsServiceInstance.GetJobs(connInfo, out jobs);
+                    if (returnValue.Succeeded)
                     {
-                        
-                        foreach (var job in this.jobs.Values)
-                        {
-                            agentJobs.Add(JobUtilities.ConvertToAgentJobInfo(job));
-                        }
+                        result.Succeeded = true;
+                        result.Jobs = jobs.ToArray();  
                     }
-                    result.Succeeded = true;
-                    result.Jobs = agentJobs.ToArray();
+                    else
+                    {
+                        result.ErrorMessage = returnValue.ErrorMessage;
+                    }
                 }
-                
+
                 await requestContext.SendResult(result);
-            }
-            catch (Exception e)
-            {
-                await requestContext.SendError(e);
-            }
+            });            
         }
 
         /// <summary>
@@ -133,63 +148,37 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
         /// </summary>
         internal async Task HandleJobHistoryRequest(AgentJobHistoryParams parameters, RequestContext<AgentJobHistoryResult> requestContext) 
         {
-            try 
+            await Task.Run(async () =>
             {
-                var result = new AgentJobHistoryResult();
-                ConnectionInfo connInfo;
-                ConnectionServiceInstance.TryFindConnection(
-                    parameters.OwnerUri,
-                    out connInfo);
-                if (connInfo != null)
+                try 
                 {
-                    Tuple<SqlConnectionInfo, DataTable> tuple = CreateSqlConnection(connInfo, parameters.JobId);
-                    SqlConnectionInfo sqlConnInfo = tuple.Item1;
-                    DataTable dt = tuple.Item2;
-                    int count = dt.Rows.Count;
-                    var agentJobs = new List<AgentJobHistoryInfo>();
-
-                    var agentStepMap = new Dictionary<DateTime, List<AgentJobStep>>();
-                    for (int i = 0; i < count; ++i)
+                    var result = new AgentJobHistoryResult();
+                    ConnectionInfo connInfo;
+                    ConnectionServiceInstance.TryFindConnection(
+                        parameters.OwnerUri,
+                        out connInfo);
+                    if (connInfo != null)
                     {
-                        var job = dt.Rows[i];
-                        if (JobUtilities.IsStep(job, sqlConnInfo))
+                        List<AgentJobHistoryInfo> jobHistories;
+                        var returnValue = this.AgentJobsServiceInstance.GetJobHistory(connInfo, parameters.JobId, out jobHistories);
+                        if (returnValue.Succeeded)
                         {
-                            var agentJobStep = JobUtilities.ConvertToAgentJobStep(job, sqlConnInfo);
-                            if (agentStepMap.ContainsKey(agentJobStep.RunDate))
-                            {
-                                agentStepMap[agentJobStep.RunDate].Add(agentJobStep);
-                            }
-                            else
-                            {
-                                var agentJobSteps = new List<AgentJobStep>();
-                                agentJobSteps.Add(agentJobStep);
-                                agentStepMap[agentJobStep.RunDate] = agentJobSteps;
-                            }
+                            result.Succeeded = true;
+                            result.Jobs = jobHistories.ToArray();
                         }
                         else
                         {
-                            var agentJobHistoryInfo = JobUtilities.ConvertToAgentJobHistoryInfo(job, sqlConnInfo);
-                            agentJobs.Add(agentJobHistoryInfo);
+                            result.ErrorMessage = returnValue.ErrorMessage;
                         }
+
+                        await requestContext.SendResult(result);
                     }
-                    result.Succeeded = true;
-                    foreach (AgentJobHistoryInfo agentJobHistoryInfo in agentJobs)
-                    {
-                        if (agentStepMap.ContainsKey(agentJobHistoryInfo.RunDate))
-                        { 
-                            var agentStepList = agentStepMap[agentJobHistoryInfo.RunDate].ToList();
-                            agentStepList.Sort(delegate (AgentJobStep s1, AgentJobStep s2) { return s1.StepId.CompareTo(s2.StepId); });
-                            agentJobHistoryInfo.Steps = agentStepList.ToArray();
-                        }
-                    }
-                    result.Jobs = agentJobs.ToArray();
-                    await requestContext.SendResult(result);
                 }
-            }
-            catch (Exception e) 
-            {
-                await requestContext.SendError(e);
-            }
+                catch (Exception e) 
+                {
+                    await requestContext.SendError(e);
+                }
+            });
         }
 
         /// <summary>
@@ -197,62 +186,52 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
         /// </summary>
         internal async Task HandleJobActionRequest(AgentJobActionParams parameters, RequestContext<AgentJobActionResult> requestContext)
         {
-            var result = new AgentJobActionResult();
-            try 
+            await Task.Run(async () =>
             {
+                var result = new AgentJobActionResult();
                 ConnectionInfo connInfo;
                 ConnectionServiceInstance.TryFindConnection(
                     parameters.OwnerUri,
                     out connInfo);
                 if (connInfo != null)
                 {
-                    var sqlConnection = ConnectionService.OpenSqlConnection(connInfo);
-                    var serverConnection = new ServerConnection(sqlConnection);     
-                    var jobHelper = new JobHelper(serverConnection);
-                    jobHelper.JobName = parameters.JobName;
-                    switch(parameters.Action)
+                    var returnValue = this.AgentJobsServiceInstance.InvokeJobAction(connInfo, parameters.Action, parameters.JobName);
+                    if (returnValue.Succeeded)
                     {
-                        case "run":
-                            jobHelper.Start();
-                            break;
-                        case "stop":
-                            jobHelper.Stop();
-                            break;
-                        case "delete":
-                            jobHelper.Delete();
-                            break;
-                        case "enable":
-                            jobHelper.Enable(true);
-                            break;
-                        case "disable":
-                            jobHelper.Enable(false);
-                            break;
-                        default:
-                            break;
+                        result.Succeeded = true;
                     }
-                    result.Succeeded = true;
-                    await requestContext.SendResult(result);
-                }
-            }
-            catch (Exception e) 
-            {
-                result.Succeeded = false;
-                result.ErrorMessage = e.Message;
-                await requestContext.SendResult(result);
-            }
+                    else
+                    {
+                        result.ErrorMessage = returnValue.ErrorMessage;
+                    }
+
+                    await requestContext.SendResult(result);              
+                }              
+            });
         }
 
-        private Tuple<SqlConnectionInfo, DataTable> CreateSqlConnection(ConnectionInfo connInfo, String jobId)
+        /// <summary>
+        /// Handle request to get Agent alerts list
+        /// </summary>        
+        internal async Task HandleAgentAlertsRequest(AgentAlertsParams parameters, RequestContext<AgentAlertsResult> requestContext)
         {
-            var sqlConnection = ConnectionService.OpenSqlConnection(connInfo);
-            var serverConnection = new ServerConnection(sqlConnection);     
-            var server = new Server(serverConnection);       
-            var filter = new JobHistoryFilter(); 
-            filter.JobID = new Guid(jobId);
-            var dt = server.JobServer.EnumJobHistory(filter);
-            var sqlConnInfo = new SqlConnectionInfo(serverConnection, SqlServer.Management.Common.ConnectionType.SqlConnection);
-            return new Tuple<SqlConnectionInfo, DataTable>(sqlConnInfo, dt);
+
+
+            await requestContext.SendResult(null);
         }
+
+
+        // private Tuple<SqlConnectionInfo, DataTable> CreateSqlConnection(ConnectionInfo connInfo, String jobId)
+        // {
+        //     var sqlConnection = ConnectionService.OpenSqlConnection(connInfo);
+        //     var serverConnection = new ServerConnection(sqlConnection);     
+        //     var server = new Server(serverConnection);       
+        //     var filter = new JobHistoryFilter(); 
+        //     filter.JobID = new Guid(jobId);
+        //     var dt = server.JobServer.EnumJobHistory(filter);
+        //     var sqlConnInfo = new SqlConnectionInfo(serverConnection, SqlServer.Management.Common.ConnectionType.SqlConnection);
+        //     return new Tuple<SqlConnectionInfo, DataTable>(sqlConnInfo, dt);
+        // }
 
     }
 }
