+++ src/Microsoft.SqlTools.ServiceLayer/Agent/Contracts/AgentAlertInfo.cs
@@ -1,0 +1,58 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+//
+
+using System;
+using System.Collections.Generic;
+using Microsoft.SqlTools.ServiceLayer.Agent;
+
+namespace Microsoft.SqlTools.ServiceLayer.Agent.Contracts
+{
+    [Flags]
+    public enum NotifyMethods
+    {
+        None = 0,
+        NotifyEmail = 1,
+        Pager = 2,
+        NetSend = 4,
+        NotifyAll = 7
+    }
+
+    public enum AlertType
+    {
+        SqlServerEvent = 1,
+        SqlServerPerformanceCondition = 2,
+        NonSqlServerEvent = 3,
+        WmiEvent = 4
+    }
+
+    /// <summary>
+    /// a class for storing various properties of agent alerts
+    /// </summary>
+    public class AgentAlertInfo
+    {
+        public int Id { get; }
+        public int DelayBetweenResponses { get; set; }
+        public string EventDescriptionKeyword { get; set; }
+        public string EventSource { get; }
+        public int HasNotification { get; }
+        public NotifyMethods IncludeEventDescription { get; set; }
+        public bool IsEnabled { get; set; }
+        public Guid JobId { get; set; }
+        public string JobName { get; }
+        public DateTime LastOccurrenceDate { get; set; }
+        public DateTime LastResponseDate { get; set; }
+        public int MessageId { get; set; }
+        public string NotificationMessage { get; set; }
+        public int OccurrenceCount { get; }
+        public string PerformanceCondition { get; set; }        
+        public int Severity { get; set; }
+        public string DatabaseName { get; set; }
+        public DateTime CountResetDate { get; }
+        public string CategoryName { get; set; }
+        public AlertType AlertType { get; }
+        public string WmiEventNamespace { get; set; }
+        public string WmiEventQuery { get; set; }
+    }
+}
